requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
python-docx==1.0.1
PyPDF2==3.0.1
python-pptx==0.6.21
supabase==2.3.0
python-dotenv==1.0.0
pydantic==2.5.2
tiktoken>=0.5.2,<0.6.0 